requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
dj-database-url
pytest
pytest-django
Faker
flake8    
black     
isort     
bcrypt
pyjwt
requests